"Create form data object before calling addPostCart in CreateCart component"
Explanation:
The code diff shows the addition of an interface `formType` and some changes in the `onSubmit` function of the `CreateCart` component. Instead of passing the form data directly to the `addPostCart` function, a `formDataObject` of type `formType` is created and then passed to the `addPostCart` function. This change ensures that the data sent to the `addPostCart` function is of the correct type and format.